bdx: Improve occur by calling `ivy-read' with an :action function
This way, the action is saved as part of `ivy-state' and we can use
that in the occur buffer.

Buttons are also added to the occur buffer, that invoke the action for
item that was clicked.

* README.md: (bdx.el): Modify doc for `bdx-query'.
* bdx.el: Modify doc in commentary.
(bdx-query): Add `action' argument, pass it to `ivy-read' and save as
extra prop.
(bdx--occur): Use that action retrieved from extra-props to add
buttons for each candidate that invoke the action.
(bdx-disassemble, bdx-find-definition): Pass self as :action.